refactoring package distributions
