--- src/main/java/UserProfile/PersonService.java
@@ -1,11 +1,0 @@
-package UserProfile;
-
-import entity.PersonRepo;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-public class PersonService {
-    @Autowired
-    private PersonRepo personRepo;
-}
