+++ src/main/java/com/spotify/spotify_web_application/SpotifyWebApplication.java
@@ -6,8 +6,9 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class SpotifyWebApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(SpotifyWebApplication.class, args);
-	}
+    public static void main(String[] args) {
+
+        SpringApplication.run(SpotifyWebApplication.class, args);
+    }
 
 }
