+++ src/main/java/com/spotify/spotify_web_application/entity/Person.java
@@ -1,0 +1,15 @@
+package com.spotify.spotify_web_application.entity;
+
+
+import jakarta.persistence.*;
+
+@Entity
+@Table
+public class Person {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "username", nullable = false, unique = true)
+    private String username;
+}
