+++ src/main/java/com/spotify/spotify_web_application/entity/PersonRepo.java
@@ -1,0 +1,6 @@
+package com.spotify.spotify_web_application.entity;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PersonRepo extends JpaRepository<Person,Long> {
+}
