+++ src/main/java/com/spotify/spotify_web_application/entity/UserDetails.java
@@ -1,0 +1,52 @@
+package com.spotify.spotify_web_application.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.io.Serializable;
+import java.util.Date;
+
+@Entity
+@Table(name="user_details")
+@Getter
+@Setter
+public class UserDetails implements Serializable {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "username", nullable = false, unique = true)
+    private String username;
+
+    @Column(name = "password", nullable = false)
+    private String password;
+
+    @Column(name = "email", nullable = false, unique = true)
+    private String email;
+
+    @Column(name = "first_name")
+    private String firstName;
+
+    @Column(name = "last_name")
+    private String lastName;
+
+    @Column(name = "token")
+    private String accessToken;
+
+    @Column(name = "refresh_token")
+    private String refreshToken;
+
+    @Column(name = "ref_id")
+    private String refId;
+
+    @Temporal(TemporalType.TIMESTAMP)
+    private Date createdOn;
+
+    @Temporal(TemporalType.TIMESTAMP)
+    private Date updatedOn;
+
+
+
+
+}
