--- src/main/java/entity/Person.java
@@ -1,15 +1,0 @@
-package entity;
-
-
-import jakarta.persistence.*;
-
-@Entity
-@Table
-public class Person {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "username", nullable = false, unique = true)
-    private String username;
-}
