--- src/main/java/entity/PersonRepo.java
@@ -1,6 +1,0 @@
-package entity;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PersonRepo extends JpaRepository<Person,Long> {
-}
