--- src/main/java/entity/UserDetailsRepository.java
@@ -1,9 +1,0 @@
-package entity;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserDetailsRepository extends JpaRepository<UserDetails, Long> {
-
-    UserDetails findByRefId(String refId);
-
-}
